Added 'data_path' property
Because multiple Pill objects can now be stored in the same directory with different prefixes, it is occasionally useful to be able to check what path a Pill is writing to.

As an example: I've created a subclass of a Boto3 session. It contains helper methods for generating new sessions via MFA or AssumeRole, and each new session also contains a reference to its parent session - sort of a one-way linked list. I'm going to include a Pill object inside the class, and ensure that child sessions inherit the same status (record/playback), path, and prefix as their parents, with their position in the session list appended to the prefix. This gives me an easy way to record and playback the permissions escalation from my base credentials to an MFA authenticated role session.