Record empty body if an object is of bytes type
During recording a S3 object retrieving request the library raises
`TypeError` while the response is totally valid Even a binary object
received by a client couldn't be recorded it shouldn't be an issue for
recording of all the rest of response. Lack of file often is not an
issue for test run. Also possible to record file. File recording is not
included to the PR because of potential security issues.