Rename variable 'lambda' README examples (#72)
`lambda` is a Python reserved word, so best practice is not to override it as it prevents its use later in the file. The syntax highlighting on GitHub is also showing it as a keyword which is a little confusing. I also commented the non-code lines so the examples are valid Python.