Record based on parameter hashing
This changes the save/load behavior from incrementing on each call
to hashing the parameters and body in the API request.

So if the same request is sent and we recieve different parameters
it should overwrite the previous response. The exception to this
is if the api call is paginated, in which case we should increment
the index counter rather then overwriting.

Behavior on pagination vs incrementing still needs testing.